+++ gg-pingpong-api/src/main/java/gg/pingpong/api/admin/match/controller/response/EnrolledMatchesResponseDto.java
@@ -1,4 +1,4 @@
-package gg.pingpong.api.admin.match.dto;
+package gg.pingpong.api.admin.match.controller.response;
 
 import java.util.List;
